Refactor OwnerDashboard by splitting into modular components
